add role into authentication

+++ passport-jwt-v2/routes/user.js
@@ -2,9 +2,10 @@ const router = require('express').Router(),
     bcrypt = require('bcrypt')
     User = require('../models/Users')
 
-// get user
+    
+// get user profile
 router.get('/profile',async(req,res) => {
-    return res.json(req.user)
+    return res.json(req.user).end()
 })
 
 // update user
@@ -17,13 +18,17 @@ router.put('/profile',async(req,res) => {
         lastName,
         birthDate
     }})
-    return res.json({"message":"Update success"})
+    return res.json({"message":"Update success"}).end()
 })
 
 // delete user
 router.delete('/profile',async(req,res) => {
-    const user = await User.findOneAndDelete({"username":req.user.username})
-    return res.json({"message":"Delete success"})
+    if (req.user.role == "admin"){
+        const user = await User.findOneAndDelete({"username":req.user.username})
+        return res.json({"message":"Delete success"}).end()
+    }else{
+        return res.status(403).json({"status": 403,"message": "Forbidden"}).end()
+    }
 })
 
 module.exports = router;
